requirements.txt
streamlit
st-link-analysis
streamlit-community-navigation-bar
authlib
futures
azure-storage-blob
gremlinpython
matplotlib
networkx
pyvis
stvis